requirements.txt
python-dotenv>=1.0.1
agno>=0.6.0
groq>=0.11.0

